tests: Move ivshmem to integration_ivshmem.rs
Extract mod ivshmem from integration.rs into a standalone
test binary integration_ivshmem.rs. Move shared live
migration helpers (start_live_migration, print_and_panic)
into common/live_migration.rs so they can be reused by
both integration_live_migration.rs and the new file.
